+++ pyshacl/constraints/sparql/sparql_based_constraints.py
@@ -371,13 +371,19 @@ class SPARQLBasedConstraint(ConstraintComponent):
 
             except ValidationFailure as e:
                 raise e
+            if not self.shape.is_property_shape:
+                result_val = f
+            else:
+                result_val = None
             for v in violating_vals:
                 non_conformant = True
                 if isinstance(v, bool) and v is True:
                     rept = self.make_v_result(
-                        f, **rept_kwargs)
+                        f, value_node=result_val, **rept_kwargs)
                 elif isinstance(v, tuple):
                     t, p, v = v
+                    if v is None:
+                        v = result_val
                     rept = self.make_v_result(
                         t or f, value_node=v, result_path=p,
                         **rept_kwargs)